requirements.txt
decorator==4.0.6
Django==1.9
django-debug-toolbar==1.4
django-extensions==1.6.1
djangorestframework==3.3.2
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.1.18
traitlets==4.0.0
Werkzeug==0.11.3
wheel==0.24.0
